Move MPI and CUDA to Julia extensions
