+++ Project.toml
@@ -8,13 +8,17 @@ authors = [
     "Jake Bolewski <clima-software@caltech.edu>",
     "Gabriele Bozzola <gbozzola@caltech.edu>",
 ]
-version = "0.5.8"
+version = "0.5.9"
 
-[deps]
+[weakdeps]
 CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
+
+[extensions]
+ClimaCommsCUDAExt = "CUDA"
+ClimaCommsMPIExt = "MPI"
 
 [compat]
 CUDA = "3, 4, 5"
 MPI = "0.20.18"
-julia = "1.8"
+julia = "1.9"
